requirements.txt
opencv-contrib-python
numpy
ultralytics
Pillow
